removing pipefail
This will prevent issues similar to the problem we had with the staging cluster which had somehow ended up with its metrics api version getting deleted.  We still want as much data as possible to be pulled from any queries that ran fine.  The failure will still be visible to the person running the script.